# Optimize Eq and Hash for Path/PathBuf