@the8472 2021-11-05
Almost entirely improvements except for `webrender-wrench` which spends more time in llvm. I guess that it makes use of a lot of paths and has to spend more time to optimize it, likely trading compile time for runtime in that case.