@joshtriplett 2021-11-07
This seems like a reasonable set of performance numbers to me; it's paying compile time to get more runtime performance.