@Mark-Simulacrum 2021-11-09
OK, this looks good to me. I think the commit history can be cleaned up a little -- 8 commits seems a little too many :)

But otherwise r=me.