@rust-log-analyzer 2021-11-10
The job **`i686-msvc-1`** failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/runs/4160008743?check_suite_focus=true) [(plain)](https://github.com/rust-lang-ci/rust/commit/2ec3ee22b5808ab9a312c46fff596a7789f58752/checks/4160008743/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

failures:

---- path::tests::test_compare stdout ----
thread 'path::tests::test_compare' panicked at '"c:\\foo" == "C:\\foo", expected true, got 16285471172997689862 and 1937321254420185683', library\std\src\path\tests.rs:1538:9

failures:
    path::tests::test_compare


test result: FAILED. 864 passed; 1 failed; 3 ignored; 0 measured; 0 filtered out; finished in 14.84s

error: test failed, to rerun pass '-p std --lib'


command did not execute successfully: "\\\\?\\D:\\a\\rust\\rust\\build\\i686-pc-windows-msvc\\stage0\\bin\\cargo.exe" "test" "--target" "i686-pc-windows-msvc" "-Zbinary-dep-depinfo" "-j" "8" "--release" "--locked" "--color" "always" "--features" "panic-unwind backtrace compiler-builtins-c" "--manifest-path" "D:\\a\\rust\\rust\\library/test/Cargo.toml" "-p" "std" "--"


Build completed unsuccessfully in 0:50:40
Build completed unsuccessfully in 0:50:40
make: *** [Makefile:72: ci-subset-1] Error 1
```

</details>