@the8472 2021-11-10
Apparently the drive prefixes are the *only* thing where we do case-insensitive comparison on paths. 😮‍💨
Not sure yet how to solve that efficiently.