@ChrisDenton 2021-11-10
> Apparently the drive prefixes are the _only_ thing where we do case-insensitive comparison on paths.
They are the only place that are guaranteed to be ASCII case-insensitive. Other parts of the path could be case-sensitive.