@the8472 2021-11-11
@Mark-Simulacrum ready for another review, the latest commit should fix the windows failure.